+++ src/main/services/base/Platform.js
